pip_requirements.txt
pyparsing==2.4
simplejson==3.17
requests==2.25
fusepy>=3.0.1
psycopg2==2.9
Django==3.2
django-cors-headers==3.7
django-extensions==3.1.3
django-debug-toolbar==3.2.1
django-tastypie==0.14.3
# Use our version of django-tastypie-swagger, the official one has been abandoned
git+https://github.com/bmampaey/django-tastypie-swagger.git
swagger-spec-validator==2.7.3
coverage
ipython
